+++ config/wdio/wdio.conf.js
@@ -1,4 +1,5 @@
 const localIP = require('ip');
+const glob = require('glob');
 
 const AxeService = require('../../lib/wdio/services').Axe;
 const TerraService = require('../../lib/wdio/services').Terra;
@@ -6,11 +7,14 @@ const SeleniumDockerService = require('../../lib/wdio/services').SeleniumDocker;
 const visualRegressionConfig = require('../../lib/wdio/visualRegressionConf');
 const ServeStaticService = require('../../lib/wdio/services/index').ServeStaticService;
 const path = require('path');
+const PackageUtilities = require('lerna/lib/PackageUtilities');
+const Repository = require('lerna/lib/Repository');
 
 const ip = process.env.WDIO_EXTERNAL_HOST || localIP.address();
 const webpackPort = process.env.WDIO_EXTERNAL_PORT || 8080;
+const ci = process.env.TRAVIS || process.env.CI;
 
-exports.config = {
+const config = {
   specs: [
     path.join('tests', 'wdio', '**', '*-spec.js'),
     path.join('packages', '*', 'tests', 'wdio', '**', '*-spec.js'),
@@ -25,7 +29,7 @@ exports.config = {
   sync: true,
   logLevel: 'silent',
   coloredLogs: true,
-  bail: 0,
+  bail: ci ? 1 : 0,
   screenshotPath: path.join('.', 'errorScreenshots'),
   waitforTimeout: 3000,
   connectionRetryTimeout: 90000,
@@ -35,6 +39,10 @@ exports.config = {
   visualRegression: visualRegressionConfig,
 
   baseUrl: `http://${ip}:${webpackPort}`,
+
+  seleniumDocker: {
+    enabled: !ci,
+  },
 
   // Ignore deprecation warnings. When chrome supports /actions API we'll update to use those.
   deprecationWarnings: false,
@@ -47,5 +55,33 @@ exports.config = {
   mochaOpts: {
     ui: 'bdd',
     timeout: 20000,
+    bail: ci,
   },
 };
+
+if (ci) {
+  config.host = 'standalone-chrome';
+}
+
+// This code only executes for monorepos.  It will create a set of suites that can then be executed
+// independently and/or in parallel via 'wdio --suite suite1' for example
+const isRepoTest = !process.cwd().includes('packages');
+if (isRepoTest) {
+  const packageLocationsWithTests = PackageUtilities.getPackages(new Repository(path.resolve('.')))
+    // eslint-disable-next-line no-underscore-dangle
+    .map(pkg => path.join(pkg._location, 'tests', 'wdio', '**', '*-spec.js'))
+    .filter(packageLocation => glob.sync(packageLocation).length > 0);
+
+  const numberOfSuites = 4;
+  config.suites = {};
+  [...Array(numberOfSuites)].forEach((_, index) => {
+    config.suites[`suite${index + 1}`] = [];
+  });
+  const itemsPerSuite = Math.ceil(packageLocationsWithTests.length / numberOfSuites);
+  packageLocationsWithTests.forEach((packageLocation, index) => {
+    const currentSuite = `suite${Math.floor(index / itemsPerSuite) + 1}`;
+    config.suites[currentSuite] = config.suites[currentSuite].concat(packageLocation);
+  });
+}
+
+exports.config = config;