+++ scripts/serve/serve-cli.js
@@ -9,6 +9,7 @@ const loadDefaultWebpackConfig = require('./loadDefaultWebpackConfig');
 commander
   .version(packageJson.version)
   .option('--config <path>', 'The webpack config to serve.', undefined)
+  .option('--host <s>', 'Sets the host that the server will listen on. eg. \'10.10.10.1\'', undefined)
   .option('--port <n>', 'The port the server should listen on.', parseInt)
   .option('-p, --production', 'Passes the -p flag to the webpack config')
   .parse(process.argv);
@@ -17,6 +18,7 @@ const port = commander.port || process.env.PORT;
 
 serve({
   config: loadDefaultWebpackConfig(commander.config),
+  host: commander.host,
   port,
   production: commander.production,
 });